move CustomTransactionError to subtensor common
