+++ pallets/subtensor/src/lib.rs
@@ -21,7 +21,7 @@ use pallet_balances::Call as BalancesCall;
 // use pallet_scheduler as Scheduler;
 use scale_info::TypeInfo;
 use sp_core::Get;
-use sp_runtime::{DispatchError, transaction_validity::TransactionValidityError};
+use sp_runtime::DispatchError;
 use sp_std::marker::PhantomData;
 use subtensor_runtime_common::{AlphaCurrency, Currency, CurrencyReserve, NetUid, TaoCurrency};
 
@@ -2436,67 +2436,6 @@ pub mod pallet {
             );
             Ok(())
         }
-    }
-}
-
-#[derive(Debug, PartialEq)]
-pub enum CustomTransactionError {
-    ColdkeyInSwapSchedule,
-    StakeAmountTooLow,
-    BalanceTooLow,
-    SubnetNotExists,
-    HotkeyAccountDoesntExist,
-    NotEnoughStakeToWithdraw,
-    RateLimitExceeded,
-    InsufficientLiquidity,
-    SlippageTooHigh,
-    TransferDisallowed,
-    HotKeyNotRegisteredInNetwork,
-    InvalidIpAddress,
-    ServingRateLimitExceeded,
-    InvalidPort,
-    BadRequest,
-    ZeroMaxAmount,
-    InvalidRevealRound,
-    CommitNotFound,
-    CommitBlockNotInRevealRange,
-    InputLengthsUnequal,
-    UidNotFound,
-    EvmKeyAssociateRateLimitExceeded,
-}
-
-impl From<CustomTransactionError> for u8 {
-    fn from(variant: CustomTransactionError) -> u8 {
-        match variant {
-            CustomTransactionError::ColdkeyInSwapSchedule => 0,
-            CustomTransactionError::StakeAmountTooLow => 1,
-            CustomTransactionError::BalanceTooLow => 2,
-            CustomTransactionError::SubnetNotExists => 3,
-            CustomTransactionError::HotkeyAccountDoesntExist => 4,
-            CustomTransactionError::NotEnoughStakeToWithdraw => 5,
-            CustomTransactionError::RateLimitExceeded => 6,
-            CustomTransactionError::InsufficientLiquidity => 7,
-            CustomTransactionError::SlippageTooHigh => 8,
-            CustomTransactionError::TransferDisallowed => 9,
-            CustomTransactionError::HotKeyNotRegisteredInNetwork => 10,
-            CustomTransactionError::InvalidIpAddress => 11,
-            CustomTransactionError::ServingRateLimitExceeded => 12,
-            CustomTransactionError::InvalidPort => 13,
-            CustomTransactionError::BadRequest => 255,
-            CustomTransactionError::ZeroMaxAmount => 14,
-            CustomTransactionError::InvalidRevealRound => 15,
-            CustomTransactionError::CommitNotFound => 16,
-            CustomTransactionError::CommitBlockNotInRevealRange => 17,
-            CustomTransactionError::InputLengthsUnequal => 18,
-            CustomTransactionError::UidNotFound => 19,
-            CustomTransactionError::EvmKeyAssociateRateLimitExceeded => 20,
-        }
-    }
-}
-
-impl From<CustomTransactionError> for TransactionValidityError {
-    fn from(variant: CustomTransactionError) -> Self {
-        TransactionValidityError::Invalid(InvalidTransaction::Custom(variant.into()))
     }
 }
